requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
PyQt5>=5.15.0
pytesseract>=0.3.10
pyzbar>=0.1.9
pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# Objektdetektering och streckkodsläsning
ultralytics>=8.0.0
